Gentle refactor of rate limiter code
Moved the middleware out into its own file. Can get away with this because the `app` isn't actually used inside the resulting `express-limiter` middleware (unless specific config is passed) so it can be disregarded and passed `null` instead, and the iddleware no longer needs to be created in a context where `app` is defined.
