+++ lib/api.js
@@ -1,13 +1,11 @@
 const api = require('@asl/service/api');
 const db = require('@asl/schema');
 
-const redis = require('redis');
-const limiter = require('express-limiter');
-
-const { NotFoundError, RateLimitedError } = require('./errors');
+const { NotFoundError } = require('./errors');
 
 const Workflow = require('./workflow/client');
 
+const rateLimiter = require('./middleware/rate-limiter');
 const errorHandler = require('./error-handler');
 
 module.exports = settings => {
@@ -18,19 +16,7 @@ module.exports = settings => {
   app.db = models;
 
   if (settings.limiter && settings.limiter.total) {
-    const client = redis.createClient(settings.redis);
-
-    app.use(
-      limiter(app, client)({
-        lookup: ['user.id'],
-        total: settings.limiter.total,
-        expire: 1000 * 60 * 60,
-        whitelist: req => req.path === '/me',
-        onRateLimited: function (req, res, next) {
-          next(new RateLimitedError());
-        }
-      })
-    );
+    app.use(rateLimiter(settings));
   }
 
   app.use((req, res, next) => {
